social-skills-coach/backend/requirements.txt
flask==2.2.3
flask-restful==0.3.9
flask-cors==3.0.10
flask-jwt-extended==4.4.4
passlib==1.7.4
openai>=1.0.0
requests==2.28.2
werkzeug==2.2.3
textblob==0.19.0
flask-sqlalchemy==3.0.3
psycopg2-binary==2.9.10
python-dotenv==1.0.0 